Add a dummy remove-purchase action and refactor out configuration functions into `lib/config` module.
